move connect to feature
